features: 修改精度为double

+++ SoftRasterLearning/game_math.hpp
@@ -4,12 +4,12 @@
 namespace gmath
 {
 
-	template<typename T = float> struct Vec2;
-	template<typename T = float> struct Vec3;
-	template<typename T = float> struct Vec4_hc;
+	template<typename T = double> struct Vec2;
+	template<typename T = double> struct Vec3;
+	template<typename T = double> struct Vec4_hc;
 
 
-	template<typename T = float>
+	template<typename T = double>
 	struct Vec4
 	{
 		union {
@@ -392,7 +392,7 @@ namespace gmath
 		}
 	};
 
-	template<typename T = float>
+	template<typename T = double>
 	struct Mat4x4
 	{
 		T data[16];
@@ -439,10 +439,10 @@ namespace gmath
 		static Mat4x4 Unit()
 		{
 			return Mat4x4{
-				1.f,0,0,0,
-				0,1.f,0,0,
-				0,0,1.f,0,
-				0,0,0,1.f
+				1.,0,0,0,
+				0,1.,0,0,
+				0,0,1.,0,
+				0,0,0,1.
 			};
 		}
 
@@ -450,10 +450,10 @@ namespace gmath
 		static Mat4x4 Translate(T x, T y, T z)
 		{
 			return Mat4x4{
-				1.f,0,0,x,
-				0,1.f,0,y,
-				0,0,1.f,z,
-				0,0,0,1.f
+				1.,0,0,x,
+				0,1.,0,y,
+				0,0,1.,z,
+				0,0,0,1.
 			};
 		}
 
@@ -461,20 +461,20 @@ namespace gmath
 		static Mat4x4 Rotate(T x, T y, T z)
 		{
 			return Mat4x4{
-			   1.f,0,0,0,
+			   1.,0,0,0,
 			   0, cos(x), -sin(x), 0,
 			   0, sin(x), cos(x),  0,
-			   0, 0,   0, 1.f
+			   0, 0,   0, 1.
 			} * Mat4x4{
 			   cos(y),0,sin(y),0,
-			   0,1.f,0,0,
+			   0,1.,0,0,
 			   -sin(y),0,cos(y),0,
-			   0,0,0,1.f
+			   0,0,0,1.
 			} * Mat4x4{
 			   cos(z),-sin(z),0,0,
 			   sin(z),cos(z),0,0,
-			   0,0,1.f,0,
-			   0,0,0,1.f
+			   0,0,1.,0,
+			   0,0,0,1.
 			};
 		}
 
@@ -485,7 +485,7 @@ namespace gmath
 				x, 0, 0, 0,
 				0, y, 0, 0,
 				0, 0, z, 0,
-				0, 0, 0,1.f
+				0, 0, 0,1.
 			};
 		}
